% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculateGradient}
\alias{calculateGradient}
\title{Calculate gradient}
\usage{
calculateGradient(w_t, n, design_mat, Y_col, muVector, modifyW_T = FALSE)
}
\arguments{
\item{w_t}{A vector indicating ??}

\item{n}{The number of ones to prepend to w_t}

\item{design_mat}{The design matrix}

\item{Y_col}{The column of validated Y values from the complete data matrix}

\item{muVector}{The vector calculated by calculateMu}

\item{modifyW_T}{Whether to add ones to the beginning of w_t}
}
\description{
Calculates a gradient given w_t and a design matrix
TODO
}
