% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculateHessian}
\alias{calculateHessian}
\title{Calculate Hessian Matrix}
\usage{
calculateHessian(design_mat, w_t, muVector, n, mus, modifyW_T = FALSE)
}
\arguments{
\item{design_mat}{The design matrix}

\item{w_t}{The vector ??}

\item{muVector}{The vector returned by calculateMu}

\item{n}{The number of ones to prepend to w_t}

\item{mus}{An empty, pre-allocated vector of the same length as muVector, pre-allocated memory saves time}

\item{modifyW_T}{Whether to add ones to the beginning of w_t}
}
\description{
Calculates the Hessian Matrix and lengthens w_t by n
}
