% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculateMu}
\alias{calculateMu}
\title{Calculate Mu}
\usage{
calculateMu(design_mat, prev)
}
\arguments{
\item{design_mat}{The design matrix}

\item{prev}{The previous iteration of the design matrix}
}
\description{
Calculates the value of mu according to two variables
Small helper function
}
