% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{coef.logistic2ph}
\alias{coef.logistic2ph}
\title{Coefficient Method for logistic2ph Objects}
\usage{
\method{coef}{logistic2ph}(object)
}
\arguments{
\item{object}{An object of class \code{logistic2ph}.}
}
\description{
Prints the coefficients of a \code{logistic2ph} object.
}
