% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{coefficients.linear2ph}
\alias{coefficients.linear2ph}
\title{Coefficient Method for linear2ph Objects}
\usage{
\method{coefficients}{linear2ph}(object)
}
\arguments{
\item{object}{An object of class \code{linear2ph}.}
}
\description{
Prints the coefficientss of a \code{linear2ph} object.
}
