% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.lengthenWT}
\alias{.lengthenWT}
\title{Prepend ones to a w_t}
\usage{
.lengthenWT(w_t_original, n, modifyW_T = TRUE)
}
\arguments{
\item{w_t_original}{The original vector}

\item{n}{The number of ones to add to the front of the vector}

\item{modifyW_T}{If false, instantly returns w_t_original}
}
\description{
Lengthens a vector by prepending n ones
}
