% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.pYstarCalc}
\alias{.pYstarCalc}
\title{Calculate pYstar}
\usage{
.pYstarCalc(
  gamma_design_mat,
  n,
  excludeRows,
  prev_gamma,
  comp_dat_all,
  Y_unval_index,
  pYstar,
  mu_gamma
)
}
\arguments{
\item{gamma_design_mat}{The gamma design matrix}

\item{n}{The starting row index to consider}

\item{excludeRows}{The number of rows to exclude from the first section of gamma_design_mat}

\item{prev_gamma}{The previous iteration of gamma_design_mat}

\item{comp_dat_all}{The complete dataset}

\item{Y_unval_index}{Which column of comp_dat_all houses the unvalidated Y variable}

\item{pYstar}{An empty, pre-allocated vector}

\item{mu_gamma}{An empty, pre-allocated vector}
}
\description{
TODO
}
