% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{linear2ph_class}
\alias{linear2ph_class}
\title{Constructor for linear2ph Objects}
\usage{
linear2ph_class(input)
}
\arguments{
\item{input}{the internal list input}
}
\value{
An object of class \code{linear2ph}.
}
\description{
Creates an object of class \code{linear2ph}.
}
