% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{logistic2ph_class}
\alias{logistic2ph_class}
\title{Constructor for logistic2ph Objects}
\usage{
logistic2ph_class(input)
}
\arguments{
\item{input}{the internal list input}
}
\value{
An object of class \code{logistic2ph}.
}
\description{
Creates an object of class \code{logistic2ph}.
}
