% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logreg2ph.R
\name{logreg2ph}
\alias{logreg2ph}
\title{Sieve maximum likelihood estimator (SMLE) for two-phase logistic regression problems
This function returns the sieve maximum likelihood estimators (SMLE) for the logistic regression model from Lotspeich et al. (2021)}
\usage{
logreg2ph(
  Y_unval = NULL,
  Y = NULL,
  X_unval = NULL,
  X = NULL,
  Z = NULL,
  Bspline = NULL,
  data,
  theta_pred = NULL,
  gamma_pred = NULL,
  initial_lr_params = "Zeros",
  hn_scale = 1,
  noSE = FALSE,
  TOL = 1e-04,
  MAX_ITER = 1000
)
}
\arguments{
\item{Y_unval}{Column name of the error-prone or unvalidated continuous outcome. Subjects with missing values of \code{Y_unval} are omitted from the analysis. If \code{Y_unval} is null, the outcome is assumed to be error-free.}

\item{Y}{Column name that stores the validated value of \code{Y_unval} in the second phase. Subjects with missing values of \code{Y} are considered as those not selected in the second phase. This argument is required.}

\item{X_unval}{Column name(s) with the unvalidated predictors.  If \code{X_unval} and \code{X} are \code{null}, all predictors are assumed to be error-free.}

\item{X}{Column name(s) with the validated predictors. If \code{X_unval} and \code{X} are \code{NULL}, all predictors are assumed to be error-free.}

\item{Z}{(Optional) Column name(s) with additional error-free covariates.}

\item{Bspline}{Vector of column names containing the B-spline basis functions.}

\item{data}{A dataframe with one row per subject containing columns: \code{Y_unval}, \code{Y}, \code{X_unval}, \code{X}, \code{Z}, and \code{Bspline}.}

\item{theta_pred}{Vector of columns in \code{data} that pertain to the predictors in the analysis model.}

\item{gamma_pred}{Vector of columns in \code{data} that pertain to the predictors in the outcome error model.}

\item{initial_lr_params}{Initial values for parametric model parameters. Choices include (1) \code{"Zeros"} (non-informative starting values) or (2) \code{"Complete-data"} (estimated based on validated subjects only)}

\item{hn_scale}{Size of the perturbation used in estimating the standard errors via profile likelihood. If none is supplied, default is \code{hn_scale = 1}.}

\item{noSE}{Indicator for whether standard errors are desired. Defaults to \code{noSE = FALSE}.}

\item{TOL}{Tolerance between iterations in the EM algorithm used to define convergence.}

\item{MAX_ITER}{Maximum number of iterations in the EM algorithm. The default number is \code{1000}. This argument is optional.}
}
\value{
\item{coeff}{dataframe with final coefficient and standard error estimates (where applicable) for the analysis model.}
\item{outcome_err_coeff}{dataframe with final coefficient estimates for the outcome error model.}
\item{Bspline_coeff}{dataframe with final B-spline coefficient estimates (where applicable).}
\item{vcov}{variance-covarianced matrix for \code{coeff} (where applicable).}
\item{converged}{indicator of EM algorithm convergence for parameter estimates.}
\item{se_converged}{indicator of standard error estimate convergence.}
\item{converged_msg}{(where applicable) description of non-convergence.}
\item{iterations}{number of iterations completed by EM algorithm to find parameter estimates.}
\item{od_loglik_at_conv}{value of the observed-data log-likelihood at convergence.}
}
\description{
Sieve maximum likelihood estimator (SMLE) for two-phase logistic regression problems
This function returns the sieve maximum likelihood estimators (SMLE) for the logistic regression model from Lotspeich et al. (2021)
}
\references{
Lotspeich, SC, Shepherd, BE, Amorim, GC, Shaw, PA, Tao, R. Efficient odds ratio estimation under two-phase sampling using error-prone data from a multi-national HIV research cohort. \emph{Biometrics}. 2021; 1– 12. https://doi.org/10.1111/biom.13512
}
