% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{print.linear2ph}
\alias{print.linear2ph}
\title{Print Method for linear2ph Objects}
\usage{
\method{print}{linear2ph}(object)
}
\arguments{
\item{object}{An object of class \code{linear2ph}.}
}
\description{
Prints the details of a \code{linear2ph} object.
}
