% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{print.logistic2ph}
\alias{print.logistic2ph}
\title{Print Method for logistic2ph Objects}
\usage{
\method{print}{logistic2ph}(x, ...)
}
\arguments{
\item{x}{An object of class \code{logistic2ph}.}

\item{...}{Additional arguments passed to methods}
}
\description{
Prints the details of a \code{logistic2ph} object.
}
