% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{print.summary.linear2ph}
\alias{print.summary.linear2ph}
\title{Print Method for summary.linear2ph Objects}
\usage{
\method{print}{summary.linear2ph}(x, ...)
}
\arguments{
\item{x}{An object of class \code{summary.linear2ph}.}

\item{...}{Additional arguments passed to methods}
}
\value{
Invisibly returns \code{x}.
}
\description{
Prints a structured summary of a \code{linear2ph} model.
}
