% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{print.summary.logistic2ph}
\alias{print.summary.logistic2ph}
\title{Print Method for summary.logistic2ph Objects}
\usage{
\method{print}{summary.logistic2ph}(x)
}
\arguments{
\item{x}{An object of class \code{summary.logistic2ph}.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Prints a structured summary of a \code{logistic2ph} model.
}
