% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pXCalc}
\alias{pXCalc}
\title{Calculates pX}
\usage{
pXCalc(n, comp_dat_all_cropped, errorsX, errorsY, pX, prevRows)
}
\arguments{
\item{errorsX}{If the X variable is unvalidated or could contain errors}

\item{errorsY}{If the Y variable is unvalidated or could contain errors}

\item{pX}{An empty, pre-allocated matrix}

\item{prevRows}{An empty, pre-allocated matrix}
}
\description{
TODO
}
