% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline2ph.R
\name{spline2ph}
\alias{spline2ph}
\title{Splines for two-phase regression functions}
\usage{
spline2ph(size = 20, degree = 3, data, x, group = NULL)
}
\arguments{
\item{size}{Pass on to the \code{df} argument in \code{splines::bs()}. Degrees of freedom.}

\item{degree}{Pass on to the \code{degree} argument in \code{splines::bs()}. Degree of the piecewise polynomial. Default is 3 for cubic splines.}

\item{data}{Specifies the name of the dataset. This argument is required.}

\item{x}{Column name of the covariate of the dataset.}

\item{group}{Optional. Column name of the categorical variable of which might have heterogeneous errors among different groups.}
}
\value{
the \code{data.frame} object including the original dataset and the B-spline bases.
}
\description{
Supplements splines for two-phase regression function in this package, including \code{linear2ph}, \code{logistic2ph}, \code{cv_linear2ph}, \code{cv_logistic2ph}.
}
\details{
This function can be directly applied to regression model with one error-prone covariate. If there are more than one error-prone covariate, please check vignette on how to handle it.
}
\examples{
# example code
data("mock.vccc")
spline2ph(size = 20, degree = 3, data = mock.vccc, x = "VL_unval", group = "Sex")

}
