% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{summary.logistic2ph}
\alias{summary.logistic2ph}
\title{Summary Method for logistic2ph Objects}
\usage{
\method{summary}{logistic2ph}(object)
}
\arguments{
\item{object}{An object of class \code{logistic2ph}.}
}
\value{
An object of class \code{summary.logistic2ph}, containing the call, coefficients, and covariance.
}
\description{
Summarizes the details of a \code{logistic2ph} object.
}
