% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{TwoPhase_MLE0_MEXY}
\alias{TwoPhase_MLE0_MEXY}
\title{Two Phase MLE0 MEXY}
\usage{
TwoPhase_MLE0_MEXY(Y_unval, X_unval, Y, X, Z, Bspline, hn, MAX_ITER, TOL, noSE)
}
\arguments{
\item{Y_unval}{Unvalidated Y variables}

\item{X_unval}{Unvalidated X variables}

\item{Y}{Validated Y variables}

\item{X}{Validated X variables}

\item{Z}{True covariates}

\item{Bspline}{Matrix of B splines}

\item{hn}{Scaling of hn}

\item{MAX_ITER}{Max iterations to perform when calculating convergence}

\item{TOL}{Maximum difference between iteration that satisfies convergence requirements}

\item{noSE}{Skips general spline profiling if converged}
}
\description{
TODO
}
