% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{print.logistic2ph}
\alias{print.logistic2ph}
\title{Print Method for logistic2ph Objects}
\usage{
\method{print}{logistic2ph}(object)
}
\arguments{
\item{object}{An object of class \code{logistic2ph}.}
}
\description{
Prints the details of a \code{logistic2ph} object.
}
