% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{summary.linear2ph}
\alias{summary.linear2ph}
\title{Print Method for linear2ph Objects}
\usage{
\method{summary}{linear2ph}(object)
}
\arguments{
\item{object}{An object of class \code{linear2ph}.}
}
\value{
None. Prints a summary of the object.
}
\description{
Prints the details of a \code{linear2ph} object.
}
