% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{summary.linear2ph}
\alias{summary.linear2ph}
\title{Summary Method for linear2ph Objects}
\usage{
\method{summary}{linear2ph}(object, ...)
}
\arguments{
\item{object}{An object of class \code{linear2ph}.}

\item{...}{Additional arguments passed to methods}
}
\value{
An object of class \code{summary.linear2ph}, containing the call, coefficients, and covariance.
}
\description{
Summarizes the details of a \code{linear2ph} object.
}
